resolving cmake errors
